requirements.txt
3==0.0.0
asgiref==3.2.3
beautifulsoup4==4.9.3
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0
django-appconf==1.0.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-chartjs==1.5.0
django-countries==5.5
django-crispy-forms==1.8.1
django-feather==0.2.5
django-filter==2.4.0
django-fontawesome==1.0
django-import-export==2.0
django-multi-email-field==0.6.2
django-phone-field==1.8.0
django-staticfiles==1.2.1
djangorestframework==3.10.3
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
psycopg2-binary==2.8.4
pytz==2019.3
PyYAML==5.2
six==1.13.0
soupsieve==2.1
sqlparse==0.3.0
tablib==0.14.0
virtualenv==16.7.9
whitenoise==4.1.4
xlrd==1.2.0
xlwt==1.3.0